+++ tests-integration/tests/serde_tests.rs
@@ -71,3 +71,6 @@ mod theta;
 
 #[path = "serde_tests/tuple.rs"]
 mod tuple;
+
+#[path = "serde_tests/xor.rs"]
+mod xor;